split private hpp
