+++ modules/core/src/copy.cpp
@@ -48,6 +48,7 @@
 
 #include "precomp.hpp"
 #include "opencl_kernels_core.hpp"
+#include "opencv2/core/private_iw.hpp"
 
 
 namespace cv
