+++ modules/core/src/merge.dispatch.cpp
@@ -4,6 +4,8 @@
 
 
 #include "precomp.hpp"
+#include "opencv2/core/private_iw.hpp"
+
 #include "opencl_kernels_core.hpp"
 
 #include "merge.simd.hpp"
