+++ modules/imgproc/src/histogram.cpp
@@ -46,6 +46,7 @@
 #include "opencv2/core/openvx/ovx_defs.hpp"
 
 #include "opencv2/core/utils/tls.hpp"
+#include "opencv2/core/private_iw.hpp"
 
 namespace cv
 {
