+++ modules/imgproc/src/resize.cpp
@@ -58,6 +58,7 @@
 
 #include "opencv2/core/softfloat.hpp"
 #include "fixedpoint.inl.hpp"
+#include "opencv2/core/private_iw.hpp"
 
 using namespace cv;
